Crate pancurses

Structs

Window

Enums

Input

Constants

A_BLINK

A_BOLD

A_COLOR

A_NORMAL

COLOR_BLACK

COLOR_BLUE

COLOR_CYAN

COLOR_GREEN

COLOR_MAGENTA

COLOR_RED

COLOR_WHITE

COLOR_YELLOW

ERR

KEY_F15

KEY_OFFSET

KEY_RESIZE

KEY_UNDO

OK

PDC_CHARTEXT_BITS

PDC_COLOR_SHIFT

SPECIAL_KEY_CODES

Traits

ToChtype

Functions

COLOR_PAIR

cbreak
Set cbreak mode.

curs_set
Alters the appearance of the cursor.

delwin
Deletes the window, freeing all associated memory. In the case of overlapping windows, subwindows should be deleted before the main window.

endwin
Should be called before exiting or escaping from curses mode temporarily.

flushinp
Throws away any type-ahead that has been typed by the user and has not yet been read by the program.

half_delay
Similar to cbreak(), but allows for a time limit to be specified, in tenths of a second.

has_colors
Indicates if the terminal supports, and can maniplulate color.

init_pair
Changes the definition of a color-pair.

initscr
Initialize the curses system, this must be the first function that is called.

napms
Suspends the program for the specified number of milliseconds.

newwin
Creates a new window with the given number of lines, nlines and columns, ncols.

nl
Enables the translation of a carriage return into a newline on input.

nocbreak
Set nocbreak mode.

noecho
Disables echoing typed characters.

resize_term
Attempts to resize the screen to the given size.

start_color
Initializes eight basic colors (black, red, green, yellow, blue, magenta, cyan, and white), and two global variables; COLORS and COLOR_PAIRS (respectively defining the maximum number of colors and color-pairs the terminal is capable of displaying).

use_default_colors
Allows the use of -1 as a foreground or background color with init_pair().

Type Definitions

chtype