Function pancurses::newterm

pub fn newterm(t: Option<&str>, output: FILE_p, input: FILE_p) -> SCREEN

A program that outputs to more than one terminal should use the newterm routine for each terminal instead of initscr.

A program that needs to inspect capabilities, so it can continue to run in a line-oriented mode if the terminal cannot support a screen-oriented program, would also use newterm. The routine newterm should be called once for each terminal. It returns a variable of type ScrPtr which should be saved as a reference to that terminal.

(For the PDCurses backend it's just an alternative interface for initscr(). It always returns SP, or NULL.)