Crate pancurses

Structs

Window

Constants

A_BOLD

A_NORMAL

COLOR_BLACK

COLOR_BLUE

COLOR_CYAN

COLOR_GREEN

COLOR_MAGENTA

COLOR_RED

COLOR_WHITE

COLOR_YELLOW

ERR

Functions

COLOR_PAIR

_attrset

endwin
Should be called before exiting or escaping from curses mode temporarily.

has_colors
Indicates if the terminal supports, and can maniplulate color.

init_pair
Changes the definition of a color-pair.

initscr
Initialize the curses system, this must be the first function that is called.

napms
Suspends the program for the specified number of milliseconds.

noecho
Disables echoing typed characters.

start_color
Initializes eight basic colors (black, red, green, yellow, blue, magenta, cyan, and white), and two global variables; COLORS and COLOR_PAIRS (respectively defining the maximum number of colors and color-pairs the terminal is capable of displaying).

Type Definitions

chtype