Struct pancurses::Window

pub struct Window {
    // some fields omitted
}

Methods

impl Window

fn attrset(&self, attributes: chtype) -> i32

Sets the current attributes of the given window to attributes.

fn erase(&self) -> i32

Copies blanks (i.e. the background chtype) to every cell of the window.

fn getch(&self) -> i32

Read a character from the terminal associated with the window.

In nodelay mode, if there is no input waiting, the value ERR is returned. In delay mode, the program will hang until the system passes text through to the program. Depending on the setting of cbreak(), this will be after one character or after the first newline. Unless noecho() has been set, the character will also be echoed into the designated window.

If keypad() is TRUE, and a function key is pressed, the token for that function key will be returned instead of the raw characters. If nodelay(win, TRUE) has been called on the window and no input is waiting, the value ERR is returned.

fn get_max_x(&self) -> i32

Return the maximum x value of this Window, in other words the number of columns.

fn get_max_y(&self) -> i32

Return the maximum y value of this Window, in other words the number of rows.

fn mv(&self, y: i32, x: i32) -> i32

The cursor associated with the window is moved to the given location.

This does not move the physical cursor of the terminal until refresh() is called. The position specified is relative to the upper left corner of the window, which is (0,0).

fn mvaddstr(&self, y: i32, x: i32, string: &str) -> i32

Write all the characters of the string str to the given window. The functionality is similar to calling waddch() once for each character in the string.

fn printw(&self, string: &str) -> i32

Add a string to the window at the current cursor position.

fn refresh(&self) -> i32

Copies the named window to the physical terminal screen, taking into account what is already there in order to optimize cursor movement.

This function must be called to get any output on the terminal, as other routines only manipulate data structures. Unless leaveok() has been enabled, the physical cursor of the terminal is left at the location of the window's cursor.

fn set_nodelay(&self, enabled: bool) -> i32

Controls whether wgetch() is a non-blocking call. If the option is enabled, and no input is ready, wgetch() will return ERR. If disabled, wgetch() will hang until input is ready.